Condensed mercury ips tasks and views into respective files
tasks list was getting very messy. this organization is more intuitive for the programmer and the new user. Now, for the programmer and the new user, having all the tasks available to the device in one place gives a complete sense for what is currently supported. In general a new user will have a device and wonder if exopy supports that device. This question can be answered rather quickly by checking this file.
