+++ src/labeq_exopy/tasks/tasks/instr/views/MercuryiPS_views.enaml
@@ -6,10 +6,9 @@
 #
 # The full license is in the file LICENCE, distributed with this software.
 # -----------------------------------------------------------------------------
-"""View for the SourceDCVoltageTask.
+"""Views for the Oxford Instruments MercuryiPS.
 
 """
-
 from textwrap import fill
 
 from enaml.widgets.api import (GroupBox, Label, Field, ObjectCombo)
@@ -21,6 +20,19 @@ from ...base_instr_view import InstrView
 from exopy.utils.widgets.qt_completers import QtLineCompleter
 
 
+enamldef MeasMagFieldView(InstrView): view:
+    """View for the MeasMagFieldTask.
+
+    """
+    constraints = [factory(auto_grid_layout)]
+
+    Label:
+        text = 'Wait (s)'
+    FloatField:
+        value := task.wait_time
+        tool_tip = fill("Time to wait before querying values from the "
+                        "field source.", 60)
+                        
 enamldef RampMagFieldView(InstrView): view:
     """View for the RampMagFieldTask.
 
@@ -39,3 +51,4 @@ enamldef RampMagFieldView(InstrView): view:
             text := task.ramp_val
             entries_updater << task.list_accessible_database_entries
             tool_tip = fill("Field strength set point in Teslas", 60)
+
