--- src/labeq_exopy/tasks/tasks/instr/views/meas_mag_field_view.enaml
@@ -1,33 +1,0 @@
-# -*- coding: utf-8 -*-
-# -----------------------------------------------------------------------------
-# Copyright 2015-2018 by ExopyHqcLegacy Authors, see AUTHORS for more details.
-#
-# Distributed under the terms of the BSD license.
-#
-# The full license is in the file LICENCE, distributed with this software.
-# -----------------------------------------------------------------------------
-"""View for the MeasMagFieldTask.
-
-"""
-from textwrap import fill
-
-from enaml.widgets.api import (GroupBox, Label, Field, ObjectCombo)
-from enaml.stdlib.fields import FloatField
-from enaml.layout.api import factory
-
-from labeq_exopy.utils.layouts import auto_grid_layout
-from ...base_instr_view import InstrView
-
-
-enamldef MeasMagFieldView(InstrView): view:
-    """View for the MeasMagFieldTask.
-
-    """
-    constraints = [factory(auto_grid_layout)]
-
-    Label:
-        text = 'Wait (s)'
-    FloatField:
-        value := task.wait_time
-        tool_tip = fill("Time to wait before querying values from the "
-                        "field source.", 60)
